+++ src/dvbmisc.cpp
@@ -17,7 +17,7 @@ AString JSONFormat(const AString& str)
 AString JSONTime(uint64_t dt)
 {
 	// calculate number of milliseconds between midnight 1-jan-1970 to midnight 1-jan-1980 (1972 and 1976 were leap years)
-	static const uint64_t offset = ADateTime::DaysSince1970 * 24 * 3600 * 1000;
+	static const uint64_t offset = ADateTime::DaysSince1970 * (uint64_t)24 * (uint64_t)3600 * (uint64_t)1000;
 	return AValue(dt + offset).ToString();
 }
